+++ src/pipes/index.ts
@@ -1,4 +1,5 @@
 export * from './cellContext/cellContext.pipe';
+export * from './cssGridTemplateColumns/cssGridTemplateColumns.pipe';
 export * from './dataCellContext/dataCellContext.pipe';
 export * from './readValue/readValue.pipe';
 export * from './resolveForwardRef/resolveForwardRef.pipe';